user_services and user_serialization_services implemented
