+++ src/apis/auth_api/register.py
@@ -5,8 +5,8 @@ from flask_jwt_extended import create_access_token, create_refresh_token
 
 from src.configs.development_config import db
 
-from src.services.user_services.user_creation_service import UserCreationService
-from src.services.user_services.user_serialization_service import UserSerializationService
+from src.services.user_services import UserService
+from src.services.serialization_services.user_serialization_service import UserSerializationService
 from src.schemas.auth_schemas.user_registration_schema import UserRegistrationSchema
 
 from src.decorators.auth_decorators.auth_requires_at_registration import auth_requires_at_registration
@@ -20,13 +20,13 @@ register_api_bp = Blueprint('register_api', __name__)
 @auth_requires_at_registration
 def register():
 
-    user_creation_service = UserCreationService(db)
+    user_service = UserService(db)
     user_serialization_service = UserSerializationService()
 
     user_data = g.validated_data
 
     try:
-        new_user = user_creation_service.create_user(user_data)
+        new_user = user_service.create_user(user_data)
     except ValueError as e:
         return jsonify({'message': str(e)}), 400
