+++ src/decorators/auth_decorators/auth_requires_at_login.py
@@ -4,7 +4,7 @@ from functools import wraps
 from flask import request, g
 
 from src.configs.development_config import db
-from src.services.user_services.user_authentication_service import UserAuthenticationService
+from src.services.user_services import UserService
 from src.utils.error_handling_utility.exceptions import (MethodNotAllowedException, CredentialsUnauthorizedException, MissingDataException)
 
 
@@ -14,7 +14,7 @@ def auth_requires_at_login(f):
         if request.method != 'POST':
             raise MethodNotAllowedException('Method not allowed.')
 
-        user_authentication_service = UserAuthenticationService(db)
+        user_service = UserService(db)
         user_data = request.get_json()
         if not user_data:
             raise MissingDataException('No input data provided.')
@@ -23,7 +23,7 @@ def auth_requires_at_login(f):
         password = user_data.get('password')
 
         # Check if the user logging with correct username or email
-        user = user_authentication_service.authenticate_user(username, password)
+        user = user_service.authenticate_user(username, password)
         if user is None:
             raise CredentialsUnauthorizedException('Invalid credentials, please try again.')
